+++ PayWiseBackend/PayWiseBackend/Domain/DTOs/CreateTransacaoDepositoDTO.cs
@@ -5,7 +5,7 @@ namespace PayWiseBackend.Domain.DTOs;
 public class CreateTransacaoDepositoDTO
 {
     [Required]
-    public double Valor { get; set; }
+    public decimal Valor { get; set; }
     [MaxLength(200)]
     public string? Descricao { get; set; }
 }
